requirements.txt
fastapi
uvicorn
langchain
langchain-core
langchain-community
pinecone-client
sentence-transformers
ragas
python-dotenv
pandas
datasets
